Update shared-devices.md
Added steps for Manual registration and removed the steps that appled for Preview only. @tiaraquan-  this is the seconf PR for this doc. The GA is still first week of July. I shall ping you prior to that. @andredm7 feel free to review. I added the steps to use the Shared mode toggle and removed the manual steps to add the tag.